rename files, folders, and clean up folder structure
