+++ admin/resources/js/components/PoolDashboard.tsx
@@ -7,21 +7,21 @@ import { UpdateClassification } from "./classification/UpdateClassification";
 
 import { ClassificationTableApi } from "./classification/ClassificationTable";
 import ClientProvider from "./ClientProvider";
-import CmoAssetPage from "./CmoAssetPage";
-import { CreateCmoAsset } from "./cmoAssets/CreateCmoAsset";
-import { UpdateCmoAsset } from "./cmoAssets/UpdateCmoAsset";
-import { CreateUser } from "./CreateUser";
+import CmoAssetPage from "./cmoAsset/CmoAssetPage";
+import { CreateCmoAsset } from "./cmoAsset/CreateCmoAsset";
+import { UpdateCmoAsset } from "./cmoAsset/UpdateCmoAsset";
+import { CreateUser } from "./user/CreateUser";
 import { Dashboard, exactMatch, MenuLink } from "./dashboard/Dashboard";
 import ErrorContainer from "./ErrorContainer";
 import HomePage from "./HomePage";
-import OperationalRequirementPage from "./OperationalRequirementPage";
-import { CreateOperationalRequirement } from "./operationalRequirements/CreateOperationalRequirement";
-import { UpdateOperationalRequirement } from "./operationalRequirements/UpdateOperationalRequirement";
+import OperationalRequirementPage from "./operationalRequirement/OperationalRequirementPage";
+import { CreateOperationalRequirement } from "./operationalRequirement/CreateOperationalRequirement";
+import { UpdateOperationalRequirement } from "./operationalRequirement/UpdateOperationalRequirement";
 import { CreatePoolCandidate } from "./poolCandidate/CreatePoolCandidate";
 import { UpdatePoolCandidate } from "./poolCandidate/UpdatePoolCandidate";
-import PoolCandidatePage from "./PoolCandidatePage";
-import { UpdateUser } from "./UpdateUser";
-import UserPage from "./UserPage";
+import PoolCandidatePage from "./poolCandidate/PoolCandidatePage";
+import { UpdateUser } from "./user/UpdateUser";
+import UserPage from "./user/UserPage";
 import PoolPage from "./pool/PoolPage";
 import { CreatePool } from "./pool/CreatePool";
 import { UpdatePool } from "./pool/UpdatePool";
