+++ admin/resources/js/components/cmoAsset/CmoAssetTable.tsx
@@ -1,12 +1,12 @@
 import React, { useMemo } from "react";
 import { defineMessages, useIntl } from "react-intl";
-import { GetCmoAssetsQuery, useGetCmoAssetsQuery } from "../api/generated";
-import { navigate, useLocation } from "../helpers/router";
-import { notEmpty } from "../helpers/util";
-import { FromArray } from "../types/utilityTypes";
-import commonMessages from "./commonMessages";
-import Button from "./H2Components/Button";
-import Table, { ColumnsOf } from "./Table";
+import { GetCmoAssetsQuery, useGetCmoAssetsQuery } from "../../api/generated";
+import { navigate, useLocation } from "../../helpers/router";
+import { notEmpty } from "../../helpers/util";
+import { FromArray } from "../../types/utilityTypes";
+import commonMessages from "../commonMessages";
+import Button from "../H2Components/Button";
+import Table, { ColumnsOf } from "../Table";
 
 const messages = defineMessages({
   columnIdTitle: {
