+++ admin/resources/js/components/cmoAsset/UpdateCmoAsset.tsx
