+++ admin/resources/js/components/operationalRequirement/OperationalRequirementTable.tsx
@@ -3,13 +3,13 @@ import { defineMessages, useIntl } from "react-intl";
 import {
   GetOperationalRequirementsQuery,
   useGetOperationalRequirementsQuery,
-} from "../api/generated";
-import { navigate, useLocation } from "../helpers/router";
-import { notEmpty } from "../helpers/util";
-import { FromArray } from "../types/utilityTypes";
-import commonMessages from "./commonMessages";
-import Button from "./H2Components/Button";
-import Table, { ColumnsOf } from "./Table";
+} from "../../api/generated";
+import { navigate, useLocation } from "../../helpers/router";
+import { notEmpty } from "../../helpers/util";
+import { FromArray } from "../../types/utilityTypes";
+import commonMessages from "../commonMessages";
+import Button from "../H2Components/Button";
+import Table, { ColumnsOf } from "../Table";
 
 const messages = defineMessages({
   columnIdTitle: {
