+++ admin/resources/js/components/pool/PoolTable.tsx
@@ -1,13 +1,13 @@
 import React, { useMemo } from "react";
 import { defineMessages, useIntl } from "react-intl";
-import commonMessages from "./commonMessages";
-import { GetPoolsQuery, useGetPoolsQuery } from "../api/generated";
-import { navigate, useLocation } from "../helpers/router";
-import { notEmpty } from "../helpers/util";
-import { FromArray } from "../types/utilityTypes";
-import Button from "./H2Components/Button";
-import Table, { ColumnsOf } from "./Table";
-import Pill from "./H2Components/Pill";
+import commonMessages from "../commonMessages";
+import { GetPoolsQuery, useGetPoolsQuery } from "../../api/generated";
+import { navigate, useLocation } from "../../helpers/router";
+import { notEmpty } from "../../helpers/util";
+import { FromArray } from "../../types/utilityTypes";
+import Button from "../H2Components/Button";
+import Table, { ColumnsOf } from "../Table";
+import Pill from "../H2Components/Pill";
 
 const messages = defineMessages({
   columnUniqueIdentifier: {
